Refactor extract conversation.PrintFileDialog
