+++ securedrop_client/gui/conversation/__init__.py
@@ -4,3 +4,4 @@ A conversation between a source and one or more journalists.
 # Import classes here to make possible to import them from securedrop_client.gui.conversation
 from .delete import DeleteConversationDialog  # noqa: F401
 from .export import Dialog as ExportFileDialog  # noqa: F401
+from .export import PrintDialog as PrintFileDialog  # noqa: F401
