+++ securedrop_client/gui/conversation/export/__init__.py
@@ -1,1 +1,2 @@
 from .dialog import ExportDialog as Dialog  # noqa: F401
+from .print_dialog import PrintDialog  # noqa: F401
