+++ tests/conftest.py
@@ -22,9 +22,9 @@ from securedrop_client.db import (
     Source,
     make_session_maker,
 )
-from securedrop_client.gui.conversation import ExportFileDialog
+from securedrop_client.gui.conversation import ExportFileDialog, PrintFileDialog
 from securedrop_client.gui.main import Window
-from securedrop_client.gui.widgets import ModalDialog, PrintDialog
+from securedrop_client.gui.widgets import ModalDialog
 from securedrop_client.logic import Controller
 from tests import factory
 
@@ -92,7 +92,7 @@ def print_dialog(mocker, homedir):
     controller = mocker.MagicMock(get_file=get_file)
     controller.qubes = False
 
-    dialog = PrintDialog(controller, file.uuid, "file123.jpg")
+    dialog = PrintFileDialog(controller, file.uuid, "file123.jpg")
 
     yield dialog
